requirements.txt
customtkinter==5.2.2
cx_Freeze==7.2.4
cx_Freeze==7.2.4
mysql_connector_repackaged==0.3.1
numpy==2.1.3
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==11.0.0
python-dotenv==1.0.1
